feat(data-fetchers): migrate bigwig to TS
